fix(dsg) rename interceptors
