+++ packages/amplication-data-service-generator/src/server/resource/resolver/to-many.template.ts
@@ -2,7 +2,7 @@ import * as common from "@nestjs/common";
 import * as graphql from "@nestjs/graphql";
 import * as nestAccessControl from "nest-access-control";
 // @ts-ignore
-import { FilterResultInterceptor } from "../../interceptors/filterResult.interceptor";
+import { AclFilterResponseInterceptor } from "../../interceptors/aclFilterResponse.interceptor";
 
 declare interface RELATED_ENTITY_WHERE_INPUT {}
 
@@ -27,7 +27,7 @@ declare const RELATED_ENTITY_NAME: string;
 export class Mixin {
   constructor(private readonly service: SERVICE) {}
 
-  @common.UseInterceptors(FilterResultInterceptor)
+  @common.UseInterceptors(AclFilterResponseInterceptor)
   @graphql.ResolveField(() => [RELATED_ENTITY])
   @nestAccessControl.UseRoles({
     resource: RELATED_ENTITY_NAME,
