+++ packages/amplication-data-service-generator/src/server/static/src/interceptors/aclFilterResponse.interceptor.ts
@@ -5,13 +5,12 @@ import {
   NestInterceptor,
 } from "@nestjs/common";
 import { Observable } from "rxjs";
+import { map } from "rxjs/operators";
 import { InjectRolesBuilder, RolesBuilder } from "nest-access-control";
 import { Reflector } from "@nestjs/core";
-import * as abacUtil from "../auth/abac.util";
-import { ForbiddenException } from "../errors";
 
 @Injectable()
-export class ValidateInputInterceptor implements NestInterceptor {
+export class AclFilterResponseInterceptor implements NestInterceptor {
   constructor(
     @InjectRolesBuilder() private readonly rolesBuilder: RolesBuilder,
     private readonly reflector: Reflector
@@ -23,13 +22,6 @@ export class ValidateInputInterceptor implements NestInterceptor {
       [context.getHandler(), context.getClass()]
     );
 
-    const type = context.getType();
-
-    const inputDataToValidate =
-      type === "http"
-        ? context.switchToHttp().getRequest().body
-        : context.getArgByIndex(1).data;
-
     const permission = this.rolesBuilder.permission({
       role: permissionsRoles.role,
       action: permissionsRoles.action,
@@ -37,17 +29,14 @@ export class ValidateInputInterceptor implements NestInterceptor {
       resource: permissionsRoles.resource,
     });
 
-    const invalidAttributes = abacUtil.getInvalidAttributes(
-      permission,
-      inputDataToValidate
+    return next.handle().pipe(
+      map((data) => {
+        if (Array.isArray(data)) {
+          return data.map((results: any) => permission.filter(results));
+        } else {
+          return permission.filter(data);
+        }
+      })
     );
-
-    if (invalidAttributes.length) {
-      throw new ForbiddenException(
-        "Insufficient privileges to complete the operation"
-      );
-    }
-
-    return next.handle();
   }
 }
